+++ mula/.ci/docker-compose.yml
@@ -31,6 +31,8 @@ services:
 
   ci_postgres:
     image: postgres:15
+    command:
+      ["postgres", "-c", "log_statement=all", "-c", "log_destination=stderr"]
     healthcheck:
       test: ["CMD", "gosu", "postgres", "pg_isready"]
       interval: 3s
@@ -47,7 +49,7 @@ services:
     volumes:
       - ./.ci/rabbitmq.conf:/etc/rabbitmq/rabbitmq.conf
     healthcheck:
-      test: [ "CMD", "rabbitmqctl", "status" ]
+      test: ["CMD", "rabbitmqctl", "status"]
       interval: 5s
       retries: 4
