--- mula/scheduler/alembic/versions/0008_add_task_schedule.py
@@ -1,98 +1,0 @@
-"""add_task_schedule
-
-Revision ID: 0008
-Revises: 0007
-Create Date: 2024-03-13 16:59:41.297670
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '0008'
-down_revision = '0007'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('p_items',
-    sa.Column('id', scheduler.utils.datastore.GUID(), nullable=False),
-    sa.Column('scheduler_id', sa.String(), nullable=True),
-    sa.Column('hash', sa.String(length=32), nullable=True),
-    sa.Column('priority', sa.Integer(), nullable=True),
-    sa.Column('task_id', scheduler.utils.datastore.GUID(), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('modified_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_p_items_hash'), 'p_items', ['hash'], unique=False)
-    op.create_table('task_runs',
-    sa.Column('id', scheduler.utils.datastore.GUID(), nullable=False),
-    sa.Column('scheduler_id', sa.String(), nullable=True),
-    sa.Column('type', sa.String(), nullable=True),
-    sa.Column('task_id', scheduler.utils.datastore.GUID(), nullable=False),
-    sa.Column('status', sa.Enum('PENDING', 'QUEUED', 'DISPATCHED', 'RUNNING', 'COMPLETED', 'FAILED', 'CANCELLED', name='taskstatus'), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('modified_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ondelete='SET NULL'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('task_schedules',
-    sa.Column('id', scheduler.utils.datastore.GUID(), nullable=False),
-    sa.Column('scheduler_id', sa.String(), nullable=True),
-    sa.Column('enabled', sa.Boolean(), nullable=False),
-    sa.Column('task_id', scheduler.utils.datastore.GUID(), nullable=True),
-    sa.Column('cron_expression', sa.String(), nullable=True),
-    sa.Column('deadline_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('evaluated_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('modified_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('task_id')
-    )
-    op.drop_index('ix_items_hash', table_name='items')
-    op.drop_table('items')
-    op.add_column('tasks', sa.Column('hash', sa.String(length=32), nullable=True))
-    op.add_column('tasks', sa.Column('data', postgresql.JSONB(astext_type=sa.Text()), nullable=False))
-    op.drop_index('ix_tasks_p_item_hash', table_name='tasks')
-    op.drop_column('tasks', 'created_at')
-    op.drop_column('tasks', 'modified_at')
-    op.drop_column('tasks', 'type')
-    op.drop_column('tasks', 'status')
-    op.drop_column('tasks', 'p_item')
-    op.drop_column('tasks', 'scheduler_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('tasks', sa.Column('scheduler_id', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column('tasks', sa.Column('p_item', postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=False))
-    op.add_column('tasks', sa.Column('status', postgresql.ENUM('PENDING', 'QUEUED', 'DISPATCHED', 'RUNNING', 'COMPLETED', 'FAILED', 'CANCELLED', name='taskstatus'), autoincrement=False, nullable=False))
-    op.add_column('tasks', sa.Column('type', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column('tasks', sa.Column('modified_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=False))
-    op.add_column('tasks', sa.Column('created_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=False))
-    op.create_index('ix_tasks_p_item_hash', 'tasks', [sa.text("(p_item ->> 'hash'::text)"), sa.text('created_at DESC')], unique=False)
-    op.drop_column('tasks', 'data')
-    op.drop_column('tasks', 'hash')
-    op.create_table('items',
-    sa.Column('id', sa.UUID(), autoincrement=False, nullable=False),
-    sa.Column('scheduler_id', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.Column('hash', sa.VARCHAR(length=32), autoincrement=False, nullable=True),
-    sa.Column('priority', sa.INTEGER(), autoincrement=False, nullable=True),
-    sa.Column('data', postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=False),
-    sa.Column('created_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=False),
-    sa.Column('modified_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name='items_pkey')
-    )
-    op.create_index('ix_items_hash', 'items', ['hash'], unique=False)
-    op.drop_table('task_schedules')
-    op.drop_table('task_runs')
-    op.drop_index(op.f('ix_p_items_hash'), table_name='p_items')
-    op.drop_table('p_items')
-    # ### end Alembic commands ###
