+++ mula/scheduler/models/task_run.py
@@ -1,3 +1,4 @@
+import enum
 import uuid
 from datetime import datetime, timezone
 
@@ -13,6 +14,30 @@ from scheduler.utils import GUID
 
 from .base import Base
 from .task_status import TaskStatus
+
+
+class TaskStatus(str, enum.Enum):
+    # Task has been created but not yet queued
+    PENDING = "pending"
+
+    # Task has been pushed onto queue and is ready to be picked up
+    QUEUED = "queued"
+
+    # Task has been picked up by a worker
+    DISPATCHED = "dispatched"
+
+    # Task has been picked up by a worker, and the worker indicates that it is
+    # running.
+    RUNNING = "running"
+
+    # Task has been completed
+    COMPLETED = "completed"
+
+    # Task has failed
+    FAILED = "failed"
+
+    # Task has been cancelled
+    CANCELLED = "cancelled"
 
 
 class TaskRun(BaseModel):
