--- mula/scheduler/models/task_status.py
@@ -1,25 +1,0 @@
-import enum
-
-
-class TaskStatus(str, enum.Enum):
-    # Task has been created but not yet queued
-    PENDING = "pending"
-
-    # Task has been pushed onto queue and is ready to be picked up
-    QUEUED = "queued"
-
-    # Task has been picked up by a worker
-    DISPATCHED = "dispatched"
-
-    # Task has been picked up by a worker, and the worker indicates that it is
-    # running.
-    RUNNING = "running"
-
-    # Task has been completed
-    COMPLETED = "completed"
-
-    # Task has failed
-    FAILED = "failed"
-
-    # Task has been cancelled
-    CANCELLED = "cancelled"
